bootcamp_week7/bootcamp_day3/requirements.txt
python-dotenv
tqdm
numpy
pandas
sentence-transformers
transformers
torch
faiss-cpu
pypdf
python-docx
scikit-learn
